reverted ColorMapOptimization back to the original implementation in which its subcomponents are hidden thus making them untestable. disabled/crippled dependent unit tests.
