+++ src/UnitTest/Core/ColorMap/ColorMapOptimizationOption.cpp
@@ -26,23 +26,30 @@
 
 #include "UnitTest.h"
 
-#include "Core/ColorMap/ColorMapOptimizationOption.h"
+// #include "Core/ColorMap/ColorMapOptimizationOption.h"
+
+/* TODO
+As the ColorMapOptimization subcomponents go back into hiding several lines of
+code had to commented out. Do not remove these lines, they may become useful
+again after a decision has been made about the way to make these subcomponents
+visible to UnitTest.
+*/
 
 // ----------------------------------------------------------------------------
 //
 // ----------------------------------------------------------------------------
-TEST(ColorMapOptimizationOption, Constructor)
+TEST(ColorMapOptimizationOption, DISABLED_Constructor)
 {
-    open3d::ColorMapOptimizationOption option;
+    // open3d::ColorMapOptimizationOption option;
 
-    EXPECT_FALSE(option.non_rigid_camera_coordinate_);
+    // EXPECT_FALSE(option.non_rigid_camera_coordinate_);
 
-    EXPECT_EQ(16, option.number_of_vertical_anchors_);
-    EXPECT_EQ(3, option.half_dilation_kernel_size_for_discontinuity_map_);
+    // EXPECT_EQ(16, option.number_of_vertical_anchors_);
+    // EXPECT_EQ(3, option.half_dilation_kernel_size_for_discontinuity_map_);
 
-    EXPECT_NEAR(0.316, option.non_rigid_anchor_point_weight_, unit_test::THRESHOLD_1E_6);
-    EXPECT_NEAR(300, option.maximum_iteration_, unit_test::THRESHOLD_1E_6);
-    EXPECT_NEAR(2.5, option.maximum_allowable_depth_, unit_test::THRESHOLD_1E_6);
-    EXPECT_NEAR(0.03, option.depth_threshold_for_visiblity_check_, unit_test::THRESHOLD_1E_6);
-    EXPECT_NEAR(0.1, option.depth_threshold_for_discontinuity_check_, unit_test::THRESHOLD_1E_6);
+    // EXPECT_NEAR(0.316, option.non_rigid_anchor_point_weight_, unit_test::THRESHOLD_1E_6);
+    // EXPECT_NEAR(300, option.maximum_iteration_, unit_test::THRESHOLD_1E_6);
+    // EXPECT_NEAR(2.5, option.maximum_allowable_depth_, unit_test::THRESHOLD_1E_6);
+    // EXPECT_NEAR(0.03, option.depth_threshold_for_visiblity_check_, unit_test::THRESHOLD_1E_6);
+    // EXPECT_NEAR(0.1, option.depth_threshold_for_discontinuity_check_, unit_test::THRESHOLD_1E_6);
 }
